+++ libraries/TLC5917/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2024-07-12
+- fix deviceCount in **write()**.
+- fix allocation buffer for channels in constructor.
+- update performance sketch.
+- add demo_setChannel sketch (array param).
+- rename blank to **outputEnable**.
+- update readme.md.
+- minor edits.
+
+----
+
 ## [0.1.2] - 2024-06-02
 - add **TLC5917_NORMAL_MODE** and **TLC5917_SPECIAL_MODE**
 - add **uint8_t getMode()**